find solution 1

+++ hw04-generics-collections/src/main/java/otus/home_work/hw04/IntegerComparator.java
@@ -5,7 +5,6 @@ import java.util.Comparator;
 public class IntegerComparator implements Comparator<Integer> {
     @Override
     public int compare(Integer first, Integer second) {
-//        return first - second;
-        return ((int) second - (int) first);
+        return (second - first);
     }
 }
